:art: fix the node details dialog
This also moves the error reporting to a new module.
